Minor API cleanup
- UIImage(View)+Reflection now takes CGFLoats for better Swift interop
- EX2Macros no longer checks if UI_USER_INTERFACE_IDIOM is available

+++ EX2Kit/Categories/UIKit/UIImage+Reflection.h
@@ -10,6 +10,6 @@
 
 @interface UIImage (Reflection)
 
-- (UIImage *)reflectedImageWithHeight:(NSUInteger)height;
+- (UIImage *)reflectedImageWithHeight:(CGFloat)height;
 
 @end
